feat: Add MainContent component
The MainContent component is added to display the main content of the ALX React app. It includes a paragraph with a list of cities the user loves to visit. This component will be used to showcase the main content section of the app.

Note: Removed meta information from the commit message.